Self-update composer instead of re-download and remove --dev flag
`composer` is already available in all Travis-CI installations, so, it’s
just a matter of self-update it in order to having the latest version.

The `--dev` flag in `composer install` is enabled by default, so it’s
not necessary.

Last, `composer` is installed globally in Travis-CI envs, so it not
needed to prepend `php` to commands.

Signed-off-by: Rogerio Prado de Jesus <rogeriopradoj@gmail.com>